internal: each tag vocabulary lives with its command
tags.go held both commands' vocabularies next to the generic dimension
mechanism; the changelog vocabulary (and the actions rendering it
feeds) moves to checks_changelog.go, the coverage vocabulary to
checks_coverage.go, and tags.go keeps only the mechanism.
